rename filtering.ts to search.ts; update tests
